Classes for module Popup and CommentsPopup have been completed and incorporated into Main.js
